Modify existing tests

+++ kernel/kernel-api/src/main/java/io/delta/kernel/exceptions/DomainDoesNotExistException.java
@@ -17,14 +17,14 @@ package io.delta.kernel.exceptions;
 
 import io.delta.kernel.annotation.Evolving;
 
-/**
- * Thrown when attempting to remove a domain metadata that does not exist in the read snapshot.
- */
+/** Thrown when attempting to remove a domain metadata that does not exist in the read snapshot. */
 @Evolving
 public class DomainDoesNotExistException extends KernelException {
-    public DomainDoesNotExistException(String domain, int snapshotVersion) {
-        super(String.format(
-            "Cannot remove domain metadata with identifier %s because it does not exist in the " +
-                "read snapshot at version %s", domain, snapshotVersion));
-    }
+  public DomainDoesNotExistException(String domain, long snapshotVersion) {
+    super(
+        String.format(
+            "Cannot remove domain metadata with identifier %s because it does not exist in the "
+                + "read snapshot at version %s",
+            domain, snapshotVersion));
+  }
 }
